+++ tests/_constants.py
@@ -396,7 +396,7 @@ EVENT_PRESENTATION_DATA = [
         "1",
         "0",
         "8000",
-        "hit",
+        "miss",
         "nan",
     ],
     [